requirements.txt
discord.py>=2.4.0
python-dotenv
sqlalchemy
psycopg2
paddleocr
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
paddlepaddle
opencv-python-headless
